[WIP] Move Network Result to Data layer and add wrapper for Mapper
