+++ data/src/main/java/com/example/data/mapper/MoviesDtoMapper.kt
@@ -1,17 +1,18 @@
 package com.example.data.mapper
 
-import android.content.res.Resources.NotFoundException
 import com.example.data.remote.dto.MoviesDto
 import com.example.domain.model.MovieModel
+import com.example.domain.wrapper.Mapper
 
-fun MoviesDto.toMovieModelList() : List<MovieModel> {
-    if(this.films.isEmpty()) throw NotFoundException("List of films is empty")
-    return this.films.map { filmDto ->
-        MovieModel(
-            id = filmDto.filmId,
-            name = filmDto.filmName,
-            releaseDate = filmDto.releaseDates.first().releaseDate,
-            imageUrl = filmDto.images.poster.x1.medium.filmImage
-        )
+class MoviesDtoMapper : Mapper<MoviesDto, List<MovieModel>> {
+    override fun mapFrom(from: MoviesDto): List<MovieModel> {
+        return from.films.map { filmDto ->
+            MovieModel(
+                id = filmDto.filmId,
+                name = filmDto.filmName,
+                releaseDate = filmDto.releaseDates.first().releaseDate,
+                imageUrl = filmDto.images.poster.x1.medium.filmImage
+            )
+        }
     }
 }
