+++ domain/src/main/java/com/example/domain/wrapper/Mapper.kt
@@ -1,0 +1,5 @@
+package com.example.domain.wrapper
+
+interface Mapper<F,T> {
+    fun mapFrom(from:F):T
+}
