+++ domain/src/test/java/com/example/domain/usecase/GetHomeMoviesListUseCaseTest.kt
@@ -1,7 +1,7 @@
 package com.example.domain.usecase
 
 import com.example.domain.model.MovieModel
-import com.example.domain.model.Resource
+import com.example.domain.wrapper.Resource
 import com.example.domain.repository.MovieRepository
 import io.mockk.coEvery
 import io.mockk.mockk
